fix: modfiy some part of the login page
