+++ src/assets/style/root.css
@@ -1,5 +1,6 @@
 body {
     #app {
         height: 100dvh;
+        display: grid;
     }
 }
